azure_functions/blobtrigger/python/pyblobtriggermi-to-queuemi/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
redis
async-timeout==4.0.2
azure-core==1.27.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
azure-storage-queue==12.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pywin32==306 ; sys_platform == 'win32'
redis==4.5.5
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.2